backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.43
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
google-generativeai==0.8.3
Pillow>=10.4.0
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
cryptography==43.0.3
